pyproject.toml
[project]
name = "voicetext"
version = "0.0.5"
description = "macOS menubar app: hold hotkey to record, release to transcribe and type"
requires-python = ">=3.13"
dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.13.0",
    "numpy>=1.24.0",
    "librosa>=0.11.0",
    "funasr_onnx>=0.4.1",
    "jieba>=0.42.1",
    "modelscope>=1.30.0",
    "pynput>=1.8.0",
    "mlx-whisper>=0.4.3",
    "openai-agents>=0.11.0",
    "fastembed>=0.7.0",
    "pyobjc-framework-Speech>=10.0",
    "pyobjc-framework-webkit>=12.1",
]
# NOTE: funasr-onnx declares numpy<=1.26.4 while fastembed declares numpy>=2.1.0,
# but fastembed works fine at runtime with numpy 1.26.4. The override below
# forces the resolver to accept numpy 1.26.4 for both.

[project.scripts]
voicetext = "voicetext.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voicetext"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv]
override-dependencies = ["numpy==1.26.4"]

[dependency-groups]
dev = [
    "py2app>=0.28.10",
    "pyinstaller>=6.19.0",
    "pytest>=9.0.2",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W"]
